requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
fake-useragent>=1.5.1

# Cloud Storage (Supabase)
supabase>=2.5.0

# LangChain Ingestion
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38

# Embeddings (Mistral)
mistralai==1.12.4

# Vector DB
chromadb==0.5.5
langchain-chroma==0.1.1
# Utilities
python-dotenv>=1.0.1
pydantic>=2.10.3,<3.0.0
tenacity>=8.3.0
loguru>=0.7.2
tqdm>=4.66.4
feedparser>=6.0.10

# Hybrid Search & Reranking
rank_bm25>=0.2.2
sentence-transformers>=3.0.0
numpy>=1.26.0

# Deployment
Flask>=3.0.0
gunicorn>=21.2.0
pysqlite3-binary
